requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
pygame>=2.5.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
numpy>=1.24.0
pyyaml>=6.0
